Move formatting logic into  custom SimpleCov profile for the application with rake task to merge coverages and report in lcov format for use in CI
Testing parallel tests

Update github actions to merge test coverages

Rever to 1 test processor
