+++ .github/workflows/test.yml
@@ -61,16 +61,20 @@ jobs:
         env:
           CI_SQL_PORT: ${{ job.services.mysql.ports['3306'] }}
           APP_IN_CI: true
-      - name: Upload to Coveralls
-        uses: coverallsapp/github-action@master
-        with:
-          github-token: ${{ secrets.GITHUB_TOKEN }}
-          path-to-lcov: ./coverage/lcov/SaraAlert.lcov
       - name: Run System Tests
         run: bundle exec rake test:system
         env:
           CI_SQL_PORT: ${{ job.services.mysql.ports['3306'] }}
           APP_IN_CI: true
+      - name: Merge Code Coverage
+        run: bundle exec rake coverage:report
+        env:
+          APP_IN_CI: true
+      - name: Upload to Coveralls
+        uses: coverallsapp/github-action@master
+        with:
+          github-token: ${{ secrets.GITHUB_TOKEN }}
+          path-to-lcov: ./coverage/lcov/SaraAlert.lcov
       - uses: actions/upload-artifact@v1
         if: failure()
         with:
