+++ test/helpers/patient_helper_test.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'test_helper'
+require 'test_case'
 
 class PatientHelperTest < ActiveSupport::TestCase
   include PatientHelper
