+++ test/models/integer_symptom_test.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'test_helper'
+require 'test_case'
 
 class IntegerSymptomTest < ActiveSupport::TestCase
   def setup; end
