+++ test/models/laboratory_test.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'test_helper'
+require 'test_case'
 
 class LaboratoryTest < ActiveSupport::TestCase
   # test "the truth" do
